mf/requirements.txt
psycopg2-binary
implicit==0.6.1
numpy==1.23.4
openTSNE==0.7.0
pandas==1.5.0
python-dotenv==0.21.1
scipy==1.9.1
